READ DESCRIPTION
I had to adapt Austin's classes to work with the rest of the code so
don't flip out if anything looks odd or different. I had to do this in
order to test my own code. The sorts DO NOT work right now but I will
get them working. Try to work with the code in Master. 714 512 4198 if
you have questions.